Remove Private and Generic drag operations from Blink
DragOperationPrivate and DragOperationGeneric are modeled after the
corresponding values form NSDragOperation, but they don't have any
corresponding meaning on the Web.

The first step to remove their usage from Chromium is to not using them
when creating DataTransfer objects in Blink. This also provides a fix
for a regression introduced in https://crrev.com/c/2196167, where the
Mojo serialization for a DragOperation fails if the value isn't one of
the Mojo enum values, as happens for "move", which Blink was assigning
the value DragOperationMove | DragOperationGeneric.

Bug: 1093536, 1093193.
Change-Id: I0fc2b65439b8b9ad2efbb7ba54e233a195d85b26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2239840
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#777312}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 445d331b38d5c0de696c89ca8c64f180829cc031