[no-behavior-changed] Add extension API to allow password migration
No behavior is changed. This CL introduces a new extension API to allow
Password Settings to move a user password from the device to their
account. The API is not called yet and is left with NOTIMPLEMENTED();
Future CLs will: a) add the relevant logic to PasswordManagerPresenter,
the same way it is done for most passwords JS APIs; b) call the API from
the new "Passwords on this device" page (crrev.com/c/2228159).

As a note, the method in the PasswordsPrivateDelegate takes the sender
web contents because it will be needed to retrieve the password
manager client.

Bug: 1090372
Change-Id: I76859af0ab22fa074a802d785ad7c9a0fc3fa912
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232371
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#777316}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa498efcd6d03a7290723477df1e22aab866407f