Mark COOP access reporting tests as failing.
Without the proper implementation, the test are expected to timeout in
99% of the cases or pass the provided expectations in 1% of the cases.

Because the 1% of the case do not happen on try-jobs, there is a risk
landing changes that will produce errors sometimes.

As long as the feature isn't landed, we mark them as
[ Timeout Pass Failure ]. This will avoid any future mistakes like in
bug 1093438.

TBR=clamy@chromium.org

BUG=1093438

Change-Id: Ieba2654bea89ed82bd8398a03582299c272e67d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241124
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#777333}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c4bdc319d513d46615c295593fcff8bc2188fce3