Test synchronization between WA API and compositor.
Previously, we had a regression where changes to the state of an
animation were not always picked up by the compositor. Some WPT
tests were written to verify that the compositor picks up the changes;
however, these tests are prone to flaking due to the need  to take a
screenshot while the animation is running to verify the behavior. This
patch does not address the flakiness of the ref tests, but ensures
proper unit test coverage in the event that the ref tests get disabled.
Followup patches will address flakiness in the WPT tests.

Bug: 1093812

Change-Id: Iab5c6853e86f0b86cf554b803bbb6c3ab47725c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232800
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#777467}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ceebf65973cecc835a4c4b160c08cc1c00bb2fc4