[NativeFS] Cleanup NFS RuntimeEnabledFeatures.
Collapse features that are no longer needed as separate ones into the
main NativeFileSystem feature. Also add a new LegacyNativeFileSystem
feature for API surface that is available for the duration of the origin
trial, but won't be available when the API ships (because it was
replaced with different API surface).

Change-Id: I40a2c2dd8e87b192e1e8e0c07e05f8d7a636decd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241938
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#777661}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8d5ee9d96848914fe8ee398795d3bafab7e4c27b