Just use if statements for input types.
There are only a couple dozen of these. It generates much less code
(comparison of AtomicStrings is an equality check of two memory
addresses), saving several kilobytes of binary size, requires no
dynamic allocation, and is likely faster because it can be executed
in a straightline fashion and benefits from inlining and branch
prediction.

Macros are used to mitigate the risk of an input type being
forgotten in one of the two places they are needed (construction and
normalization), and to make it easier to see by inspection that the
correspondences between type names and class names are correct.

Change-Id: I3a7916ae9c29d0ddcf2e18f36ae763847b5bccff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241857
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#777871}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 49baa6f345160827c4edc62f76abf8abbfb51105