Don't extend word-granularity beyond end of line
When in word-granulariry selection mode, shift clicking the end of the
line resulted in the first character of the next line being selected.
This stops at the end of the line (including whitespace).

When selecting single cells in a table, a tab or newline might be added
to the end of the value. This trims it if it's the last cell copied
from the table.

Bug: 1092554
Bug: 892750
Change-Id: I3db47cf98bba995c276b2943008986a06b1b856f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2239407
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#778083}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b40976d4c9dd469dc65be6b27a0b1ceff6fe7828