Remove all the not-site-per-process tests that aren't skipped.
Keep only the tests being skipped in NeverFixTests, which are listed
in the same order in VirtualTestSuites now, and the 2 printing tests
which are disabled in TestExpectations.

The other tests are either disabled on main because they are
fundamentally flaky (and are flaky in not-site-per-process too!) or
are enabled on main and not needed anymore in the virtual suite.

R=lukasza@chromium.org

Bug: 1069111
Change-Id: Ibac3f21d1a5387b6d370bd3a2f311d46d05aa334
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2239896
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#778423}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2fa55a8e20cf5a44d0f30e4cd863ba2d910280ab