CookieStore: GetAllForUrl to return CookieWithAccessResult
This changes GetAllForUrl to return CookieWithAccessResult which
includes CookieEffectivesSameSite data. This is so that it can
eventually switch to use EffectiveSameSite for for CookieListItem
SameSite which is mostly the same as SameSite except for when it
is UNSPECIFIED in which case it will check for the
SameSiteByDefaultCookies setting and whether the cookie is recently
created. This is per github discussion here [1]. It is based off
a change this parent change here [2], which returns
CookieWithAccessResult data from GetCookieList. The switch to have
EffectiveSameSite be returned to the client will be implemented in
a future CL tracked here [3] once all endpoints can consistently return
CookieEffectivesSameSite.

[1]https://github.com/WICG/cookie-store/issues/102
[2]https://crrev.com/c/2212697
[3]https://crbug.com/1092695

Bug: 1065010
Change-Id: Id5aa68770491506b62443db985d6594ba70bb6f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2208410
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#778485}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b9f3e6e48fac20b1107723026a5c2ce0a7585207