Protect against RemoteFrames being created with an empty frame token
As reported in crbug.com/1094850, if something goes wrong during the
creation of an HTML portal element, the frame token being used for the
RemoteFrame object associated to it could end up being an empty token.

This could result in a crash if we still use that empty token to store
references to RemoteFrame objects in the statically allocated hashmap
in remote_frame.cc, since multiple errors creating portals would lead
to the same token being used as key, resulting in an "Inserting a
duplicate item" error at runtime.

While the actual fix would be to make sure that no RemoteFrame gets
ever created when an error occurred while creating a portal (which is
the purpose of crbug.com/1094850), we can still protect our code here
so that we at least don't crash when an empty token is passed to the
RemoteFrame constructor, which preserves the original behaviour before
crrev.com/c/2172691 landed, so let's do that for now.

Bug: 1094576
Change-Id: Ie65a6402c399e225477408465f56fc19103d25d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245905
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#778708}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a69aebb420dbafa705f88c8156107ae9cc9f1cdb