Refresh sources of mutable stylesheets in getMatchedStylesForNode
Styles added to the page after DevTools was opened were not editable
because the editable source text was populated only when an
InspectorStylesheet instance was created (which happens on DevTools
startup). This CL adds synchronization of the source text for mutable
stylesheets in InspectorCSSAgent::getMatchedStylesForNode to make sure
the latest styles are available to the frontend. Also, it fixes
an issue in MergeCSSOMRulesWithText leading to incorrect merge results.

Bug: 946975
Change-Id: If558b31d042e3c7ac9d28ef8ffb896573ca13d93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238832
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#778742}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f06a240987a16a8f5c12fc8adad22c25b17fee2c