RasterModeHint should not affect restore logic
With the logic that preserves acceleration regardless of the
RasterModeHint given in ShouldAccelerate, GetOrCreateResourceProvider
will try to create a gpu resource provider even if the hint is
kPreferCPU, in which case re-creation should try to happen through
Restore() first. The hint should not have an effect on whether or not
to use Restore().

Bug: 1090180
Change-Id: Ia87b081ef611021a4c94a7ac0232c380fca3a143
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246788
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Teresa Kang <teresakang@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#778976}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d9c3390737ddec140c3b4aae62c8f1c2f4b6435