Remove PaintLayerScrollableArea compositing query assert disablers
Some of the assert disablers were obsolete (e.g. SetHas*Scrollbar) and
others are able to be scoped to non-CAP codepaths. The added disablers
in InvalidateViewportConstrainedObjects are now necessary since the
entirety of PaintLayerScrollableArea::UpdateScrollOffset is not longer
covered. These disablers are specific to non-CAP since we don't have
paint invalidation containers in CAP.

For the GraphicsLayerFor* methods, I considered forcing callers to check
for CAP (and DCHECK'ing that CAP is disabled) but the changes started
ballooning in ways that made it difficult to reason about.

Bug: 1007989

Change-Id: If624e3e5a5a699dc421ce26a2bb80501c11ea7a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247342
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#779028}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 258865ba9c447173cffa6a9c7982f6ed1a1153c3