Fixes empty alt text on broken images overriding css dimensions.
Prior to this CL, an img with CSS height, width and empty alt text
renders as nothing. This is wrong because [2] states that the img
should be rendered as a replaced element with its intrinsic dimensions.

The fix continues from the previous fix to the img title attribute by
adding an additional check to see if the alt text is empty too.

[1] https://html.spec.whatwg.org/multipage/rendering.html#images-3::represents-2

Bug: 753868
Change-Id: I9b127c54d9d5eb0f1e900dee47f24aff09f2ed8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238327
Commit-Queue: Yu Han <yuzhehan@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#779151}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2d3ccbfc625728ee69cea87257317546f942ea3b