Don't use cached ComputedStyle when flat-tree flag differs.
If we cache a ComputedStyle which is computed outside of the flat tree,
any pending resources will stay StylePendingImage. That means we cannot
use such a ComputedStyle for cached styles for a rendered element.

We already did this check for display:none subtree ComputedStyles, but
we need the same check for IsEnsuredOutsideFlatTree().

Bug: 1095669
Change-Id: Ide7395207a6d73b78bc793e4ee4c10cf31b9696f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250201
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#779530}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ecc0ab42e8581867054cba6a9b49bf59f32e6cac