Don't show scrolling contents layers that are also backface-invisible.
In this situation, we have an element like this:

<div id=target style="transform: ...; backface-visibility: hidden;
            overflow: scroll">
</div>

It will have two transform and one for scrolling:

Transform
   ScrollTranslation

However, overflow clipping is also a grouping property [1], so the
ScrollTranslation will have flattens_inherited_transform true
regardless of the transform-style style of #target.

With TransformInterop on, definitely the scrolling contents should
not be shown (except for pseudo- and stacking context descendants).
This CL changes the behavior for non-TransformInterop also, because
it doesn't seem reasonable to have the behavior the way it is right
now.

[1] https://drafts.csswg.org/css-transforms-2/#grouping-property-values

Bug: 363609

Change-Id: Ib1f356e80644582120e37261d368f774cbcaa6e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246437
Auto-Submit: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#779637}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5b5d843356d6961cc8b61f8455ea34dd8fc64d2