[FragmentItem] Fix hit-testing floats in legacy root
This patch fixes hit-testing floating objects in a legacy
root.

With |NGPaintFragment|, floats are in children, and that
regular traversal can cover this case.

With |FragmentItem|, floats are either in |NGFragmentItems|
if it has an inline formatting context, or in children
otherwise. This difference requires additional branch to the
legacy fallback when hit-testing floating objects.

Bug: 1096259
Change-Id: I8565eb663781318233c21482270804cf254e0aec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249249
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#779715}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b550067b8ecefb4478cd713f35b894876f1be360