Change NGLineTruncator to work on NGLogicalLineItem
|NGLineTruncator| was originally designed to compute and
trunate |NGPhysicalTextFragment|. This patch changes it to
work on |NGLogicalLineItem| instead.

This is yet another effort to reduce the usage of
|NGPhysicalTextFragment|, following r778837 and r779021.
This saves creating intermediate |NGPhysicalTextFragment|
when FragmentItem is enabled, and reduces dependency on
|NGPhysicalTextFragment| for the removal when the switch
to FragmentItem was completed.

Bug: 982194
Change-Id: I3742fa1ea33648e2151f7d3cbf83e977928c6715
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2230583
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#779756}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 330aad685b40a04b358523b8fa9f1ef157b444ab