Remove DisableCompositingQueryAsserts in PaintLayer
In CAP, compositing state/group mapping is not known until after
painting, so this information can't be used to isolate NeedsRepaint
operations in  MarkCompositingContainerChainForNeedsRepaint.
Scope the querying to a non-CAP codepath.

This was the last assert disabler in CAP codepaths, so add a DCHECK
to the DisableCompositingQueryAsserts constructor so no further ones
are added.

Bug: 1007989

Change-Id: I72e4a82ccf856c9ace3f3714d5e3bd0c6ccc7b86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247322
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#779879}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b0f5dcaae8fc0ea45c7236bc1f6d7e2e93b7d08