[LargestContentfulPaint] Add UMA and UKM for  experimental LCP
This CL adds UMAs and UKMs for the experimental LCP metric, which
considers removed content as valid LCP candidates. The nontrivial logic
to compute the frame merges and image-text merges is done by the
LargestContentfulPaintHandler, so the method parameters are changed to
not explicitly query the non-experimental LCP versions, and the class is
reused to computed the same merges for the experimental version.

UKM collection review:
https://docs.google.com/document/d/1Fdsg8t1d2X3nsE0QiVYnzslBjFA61NGhx0yMJQkZSEo/edit#heading=h.k5jx6iluw4yt

Bug: 1045640
Change-Id: I7d733de0bbdd12c66256b203d735efe2fa825a48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241943
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#779931}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9acc70c79105f5409af4f8fdba2f684702bc7e7c