Prevent poster rendering over video if LoadTimerFired runs after Play
If HTMLMediaElement::Play is called before
HTMLMediaElement::LoadTimerFired is run, then LoadTimerFired will
set the 'show-poster-flag' to 'true', and nothing will unset it even
after playback begins. This was because the call to
'SetShowPosterFlag(true)' was in the wrong location, this CL moves it to
the right location, and adds a test to prevent regression.

Bug: 1094778, 1095320
Change-Id: Ia1c558a35f2816136994c0013061d2e00ae90ff4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249459
Commit-Queue: Will Cassella <cassew@google.com>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#780338}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: faa9139579c34f4428719c31ca880e538224b764