Fold and remove LocalFrameView::Convert{From,To}LayoutObject
These are just wrapping calls to LocalToAbsolute* or AbsoluteToLocal*
on the LayoutObject it's passed, and does not explicitly involve the
object it is called on. Replace the usage with direct calls to the
methods on LayoutObject instead.

Bug: 634143
Change-Id: I0c2c52339f0bc7b1a321a4b9d8435153a66dc6bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250005
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#780753}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f8ed7af4656051c365421997146c8cf0806494e6