Stops marking objects that raise events that don't modify them as dirty
Some events indicate a change that does not affect the data of the accessibility node they are raised on.
For example, if a kTextSelectionChange event fires on a non-native text field,
the data on and inside the field does not need to be reserialized, since the selection endpoints are stored on the document object,
and not on the non-native text field itself.

Marking objects as dirty for any event that could be raised on them, especially
for objects with a lot of data, e.g. a long name or value attribute that is computed from descendants,
has a significant effect on performance: in the range of 100 MS for a long document
in Google Docs.

AX-Relnotes: Noticeably faster caret navigation in Google Docs
and presumably in other online editors.

R=dmazzoni@chromium.org, aleventhal@chromium.org

Change-Id: Ie138e516156951949a5f8dc272ba67dd0df1f79a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2250257
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#780780}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef7c317e707b8f023a49b85b273698c9589789be