[QuicTransport] Implement options argument
Permit an "options" argument to be supplied to the QuicTransport
constructor providing a list of certificate fingerprints that are
permitted to be used for the handshake.

Add a unit test and browser test for the new feature.

BUG=1011392

Change-Id: Ia98ab05d0d5e6fc34ae57472119f1444dab7b330
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2251869
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#780842}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dfa2b44b5fbae0f190f88b483914b9ca4a1a4900