Reland "[LayoutNG] Delay layout of positioned nodes during fragmentation"
This is a reland of d877aabea65f55906f5ae1680e0b5642dfe1a2d6

Original change :
https://chromium-review.googlesource.com/c/chromium/src/+/2213727

Revert of original change :
https://chromium-review.googlesource.com/c/chromium/src/+/2245593

-----

The original change was reverted due to web test crashes observed on
linux-webkit-msan-rel. The reason for these crashes was that
default_containing_block_.direction was not initialized, and yet
was used by LayoutFragmentainerDescendant().

To fix this, use the TextDirection of the containing_block_fragment
as the default direction in this case rather than
default_containing_block_.direction.

This also required updating GetContainingBlockInfo() to use the
TextDirection of containing_block_fragment, as well. A follow-up
change is in-progress to complete the rest of logic needed to
get the correct ContainingBlockInfo:
https://chromium-review.googlesource.com/c/chromium/src/+/2248619/2.

-----

Original change's description:
> [LayoutNG] Delay layout of positioned nodes during fragmentation
>
> When fragmenting positioned nodes, instead of placing them as soon as
> they reach their containing block, continue to bubble the positioned
> nodes up to the fragmentation context root. There, they will be placed
> and properly fragmented.
>
> This change does not implement the actual layout algorithm for
> fragmented positioned nodes. As a result, this will cause test
> failures.
>
> Bug: 1079031
> Change-Id: I3ca2f3fb669c8463ab1fc877ad510ff62526ecbf
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2213727
> Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
> Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
> Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#778177}

Bug: 1079031
Change-Id: Id0cccfbf471ae78073feb57e8dd1dd1390979b8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246550
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#780900}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0be2fce4a43210762b5f56e4d840de3bd46458a