Always cache the base computed style if there are running animations.
Previously, the base compute style was only cached if it could be used
to optimize style recalculation.  The base style is also needed for
transition retargeting.  Rather than dropping the base computed style,
it is unconditionally cached, and a flag is set to indicate if
compatible with accelerated style recalculation. Updating the cache is
deferred until after calculating animation updates. This deferral
ensures that an ElementAnimations object is attached to the element if
needed prior to the update.

Bug: 1068515, 1082401
Change-Id: Icf4bc10e08a970639ab65f30c7dab411aef82116
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252642
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#780948}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ba252773d6a33faabd771a990531689a37118752