Fixes Heap-use-after-free in ImageResourceContent::PriorityFromObservers
Prior to this CL, asynchronously applying and removing a CSS property
containing image content to an ImageInputType that's rendered as
fallback content causes heap use after use exception in
ImageResourceContent when it tries to access its ImageResourceObserver.
See minimized repro case in bug for details in the use case.
This is a regression caused by my fix [1] to bug 948066.

With the previous fix [1], the above case triggers a call into
LayoutObject [2] from ImageInputType::CreateLayoutObject() which
instead of creating a LayoutObject for its fallback content, it creates
a LayoutImage instead. This is clearly wrong. I spend some time
debugging into the lifecycle of LayoutImage, ImageResourceContent and
ImageResourceObserver but didn't track down the cause of the stale
ImageResourceObserver. Regardless, the fix is to not use
LayoutObject::CreateObject() directly, but to create layout object
subclass to render the fallback content.


[1] https://crrev.com/c/2212963
[2] https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/renderer/core/layout/layout_object.cc;l=210;drc=0a03e5eb198d0485b8cc023dbc24fd17c16c9629;bpv=0;bpt=1

Bug: 1096002
Change-Id: If059a694c9a60fb6563a446b0250532e6d19f1c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255311
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Yu Han <yuzhehan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#780992}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e7001b8845c0f229137112ea89bb486260091f29