[Text Fragment] Unflag fragment directive removal.
We want to keep the text fragment runtime enabled feature for a bit
longer since there's some holdback experiments we may want to run.
However, to unblock https://crrev.com/c/2255552 we can remove the REF
check inside the Document constructor.

This check guards the behavior where we remove the `:~:` and all
following characters from the URL's fragment on the Document and store
it in the Document's fragment_directive_ member instead. This alone will
not affect text-fragment behavior when the REF is off (text fragments
still won't work) because the fragment_directive_ is only read in
TextFragmentAnchor::TryCreateFragmentDirective which is still guarded by
the REF.

This does mean that when the REF is off that URLs will now lose the
directive part of the fragment (the :~: and after). This should be ok
for the purposes of the holdback experiment. This also affects the
TextFragmentAnchor.ElementIdFragmentFound metric which measures how
often an element id in a fragment is matched. Before this CL, if the REF
is off we would try to search for an element id like
`elementid:~:directive` which will almost certainly fail. Since we'll now
strip the directive we may find the element id in these cases. This is
consistent with what the metric is measuring and this was last used when
making changes to the API so there shouldn't be issues comparing numbers
across this change.

Bug: 919204,987027
Change-Id: I35d4df9d9c707f2d507c117d65739d3c41dca821
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2258792
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#781080}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 40ec96869bffefad335f1d94e05b41d185c7a67a