Add a Length::MinIntrinsic enum value
In the short term, we will keep this as an internal value to implement
the desired aspect-ratio behavior, where min-content will take the
aspect ratio into account but we still need to be able to access the
"real" min-content size to implement min-block-size: auto.

In the longer term, CSS may add this as a length keyword with the same
semantics and this prepares us for that.

See also https://github.com/w3c/csswg-drafts/issues/5032

Bug: 1045668
Change-Id: I76a290748e186b900bdbc2c4f5db8db55f6ce04f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255564
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#781117}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 56cccf94179e2394facb6c236c0ac45cd258414a